+++ Resources/Locale/en-US/subfloor/subfloor.ftl
@@ -1,0 +1,2 @@
+subfloor-anchor-failure = {CAPITALIZE(THE($entity))} can't be anchored here!
+subfloor-unanchor-failure = {CAPITALIZE(THE($entity))} can't be unanchored here!
